backend/requirements/common.txt
django~=3.0.0
djangorestframework
pillow
django-cors-headers
djangorestframework-jwt
django-pydenticon